autogen_lawyer/requirements.txt
autogen-agentchat[llm,gemini]~=0.2
rich
python-dotenv
autogen
